Code reorganisation and tidying
