+++ python/src/aggregator.py
@@ -1,3 +1,4 @@
+# core imports
 import logging
 import traceback
 import json
@@ -8,10 +9,16 @@ import hashlib
 from collections import Counter
 from datetime import datetime, timedelta, date
 from concurrent.futures import ThreadPoolExecutor
+
+# external library imports
 import requests
 import redis
 
+# app level imports
+from utils import seconds_to_time_string
+
 # get environment variables
+PUSH_API_NODE = os.getenv('EOSIO_PUSH_API_NODE', '')
 BLOCKS_API_NODE = os.getenv('EOSIO_BLOCKS_API_NODE', '')
 EMPTY_DB_START_BLOCK = os.getenv('EMPTY_DB_START_BLOCK', '')
 EXCLUDED_ACCOUNTS = os.getenv('EXCLUDED_ACCOUNTS','').split(',')
