+++ python/src/submitter.py
@@ -1,3 +1,4 @@
+# core imports
 import logging
 import traceback
 import json
@@ -5,9 +6,13 @@ import os
 import time
 import signal
 from datetime import datetime, timedelta, date
-from concurrent.futures import ThreadPoolExecutor
+
+# external library imports
 import requests
 import redis
+
+# app level imports
+from utils import seconds_to_time_string, get_current_data_submission_state, get_expected_dataset_id
 
 # get environment variables
 PUSH_API_NODE = os.getenv('EOSIO_PUSH_API_NODE', '')
@@ -38,41 +43,6 @@ logger.addHandler(handler)
 
 # establish connection to redis server
 redis = redis.StrictRedis(host='redis', port=6379, db=0, decode_responses=True)
-
-def seconds_to_time_string(epochsecs):
-    return datetime.fromtimestamp(epochsecs).strftime('%Y-%m-%d %H:%M')
-
-def get_current_data_submission_state():
-    '''Returns the current period_start and state (whether waiting for totals or individual account usage)'''
-    data = {"scope": CONTRACT_ACCOUNT, "code": CONTRACT_ACCOUNT, "table": 'resourceconf', "json": True}
-    data = json.dumps(data).encode("utf-8")
-    try:
-        table_info = requests.post(PUSH_API_NODE + '/v1/chain/get_table_rows', data=data, timeout=10).json()
-        logger.info(table_info)
-        rows = table_info.get('rows', [])
-        period_start = datetime.strptime(rows[0]['period_start'], '%Y-%m-%dT%H:%M:%S').timestamp()
-        inflation_transferred = bool(rows[0]['inflation_transferred'])
-        return period_start, 'INDIVIDUAL_USAGE' if inflation_transferred else 'TOTAL_USAGE'
-    except:
-        logger.info(f'get_current_data_submission_state failed')
-        logger.info(traceback.format_exc())
-        return None, ''
-
-def get_expected_dataset_id():
-    '''Returns the next dataset that is expected by the contract'''
-    data = {"scope": CONTRACT_ACCOUNT, "code": CONTRACT_ACCOUNT, "table": 'ressysusage',
-                        "lower_bound": f' {SUBMISSION_ACCOUNT}', "limit": 1, "json": True}
-    data = json.dumps(data).encode("utf-8")
-    try:
-        table_info = requests.post(PUSH_API_NODE + '/v1/chain/get_table_rows', data=data, timeout=10).json()
-        logger.info(table_info)
-        rows = table_info.get('rows', [])
-        submission_hash_list = rows[0]['submission_hash_list']
-        return len(submission_hash_list)
-    except:
-        logger.info(f'get_expected_dataset_id failed')
-        logger.info(traceback.format_exc())
-        return None
 
 # wait for redis dump file to signify that DB is ready
 while KEEP_RUNNING and not os.path.exists('/data/dump.rdb'):
